move most styling to scss, remove margin from list
