+++ cit3.0-web/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
+    "jquery": "^3.5.1",
     "leaflet": "^1.7.1",
     "node-sass": "4.14.1",
     "react": "^17.0.1",
